refactor(arbitrum): move utility functions to utils.ts
